Restructure submodules
