+++ lib/async.js
@@ -22,7 +22,4 @@ const delay = (msec, signal = null) =>
     });
   });
 
-module.exports = {
-  timeout,
-  delay,
-};
+module.exports = { timeout, delay };
