+++ lib/network.js
@@ -44,4 +44,18 @@ const fetchPolyfill = (url, options = {}) => {
   });
 };
 
-module.exports = { fetch: global.fetch || fetchPolyfill, receiveBody };
+const ipToInt = (ip = '127.0.0.1') => {
+  if (ip === '') return 0;
+  const bytes = ip.split('.');
+  let res = 0;
+  for (const byte of bytes) {
+    res = (res << 8) + parseInt(byte, 10);
+  }
+  return res;
+};
+
+module.exports = {
+  fetch: global.fetch || fetchPolyfill,
+  receiveBody,
+  ipToInt,
+};
