+++ lib/objects.js
@@ -1,13 +1,13 @@
 'use strict';
 
-const utilities = require('./utilities.js');
+const strings = require('./strings.js');
 
 const makePrivate = (instance) => {
   const iface = {};
   const fields = Object.keys(instance);
   for (const fieldName of fields) {
     const field = instance[fieldName];
-    if (utilities.isConstant(fieldName)) {
+    if (strings.isConstant(fieldName)) {
       iface[fieldName] = field;
     } else if (typeof field === 'function') {
       const bindedMethod = field.bind(instance);
@@ -37,10 +37,13 @@ const jsonParse = (buffer) => {
   }
 };
 
+const isHashObject = (o) =>
+  typeof o === 'object' && o !== null && !Array.isArray(o);
+
 const flatObject = (sourceObject, fieldNames = []) => {
   const target = {};
   for (const [key, value] of Object.entries(sourceObject)) {
-    if (!utilities.isHashObject(value)) {
+    if (!isHashObject(value)) {
       target[key] = value;
       continue;
     }
@@ -49,7 +52,7 @@ const flatObject = (sourceObject, fieldNames = []) => {
       continue;
     }
     for (const [childKey, childValue] of Object.entries(value)) {
-      const combined = `${key}${utilities.toUpperCamel(childKey)}`;
+      const combined = `${key}${strings.toUpperCamel(childKey)}`;
       if (sourceObject[combined] !== undefined) {
         const error = `Can not combine keys: key "${combined}" already exists`;
         throw new Error(error);
@@ -81,16 +84,26 @@ const unflatObject = (sourceObject, fieldNames) => {
 
 const getSignature = (method) => {
   const src = method.toString();
-  const signature = utilities.between(src, '({', '})');
+  const signature = strings.between(src, '({', '})');
   if (signature === '') return [];
   return signature.split(',').map((s) => s.trim());
+};
+
+const namespaceByPath = (namespace, path) => {
+  const [key, rest] = strings.split(path, '.');
+  const step = namespace[key];
+  if (!step) return null;
+  if (rest === '') return step;
+  return namespaceByPath(step, rest);
 };
 
 module.exports = {
   makePrivate,
   protect,
   jsonParse,
+  isHashObject,
   flatObject,
   unflatObject,
   getSignature,
+  namespaceByPath,
 };
