+++ metautil.js
@@ -1,13 +1,16 @@
 'use strict';
 
 module.exports = {
-  ...require('./lib/utilities.js'),
-  ...require('./lib/crypto.js'),
-  ...require('./lib/semaphore.js'),
-  ...require('./lib/pool.js'),
   ...require('./lib/async.js'),
-  ...require('./lib/network.js'),
+  ...require('./lib/crypto.js'),
+  ...require('./lib/datetime.js'),
   ...require('./lib/error.js'),
-  ...require('./lib/oop.js'),
   ...require('./lib/http.js'),
+  ...require('./lib/network.js'),
+  ...require('./lib/objects.js'),
+  ...require('./lib/pool.js'),
+  ...require('./lib/random.js'),
+  ...require('./lib/semaphore.js'),
+  ...require('./lib/strings.js'),
+  ...require('./lib/units.js'),
 };
