+++ test/units.js
@@ -1,0 +1,55 @@
+'use strict';
+
+const metatests = require('metatests');
+const metautil = require('..');
+
+metatests.case(
+  'Units utilities',
+  { metautil },
+  {
+    'metautil.bytesToSize': [
+      [0, '0'],
+      [1, '1'],
+      [100, '100'],
+      [999, '999'],
+      [1000, '1 KB'],
+      [1023, '1 KB'],
+      [1024, '1 KB'],
+      [1025, '1 KB'],
+      [1111, '1 KB'],
+      [2222, '2 KB'],
+      [10000, '10 KB'],
+      [1000000, '1 MB'],
+      [100000000, '100 MB'],
+      [10000000000, '10 GB'],
+      [1000000000000, '1 TB'],
+      [100000000000000, '100 TB'],
+      [10000000000000000, '10 PB'],
+      [1000000000000000000, '1 EB'],
+      [100000000000000000000, '100 EB'],
+      [10000000000000000000000, '10 ZB'],
+      [1000000000000000000000000, '1 YB'],
+    ],
+    'metautil.sizeToBytes': [
+      ['', NaN],
+      ['0', 0],
+      ['1', 1],
+      ['512', 512],
+      ['100', 100],
+      ['999', 999],
+      ['1 kB', 1000],
+      ['2 Kb', 2000],
+      ['10 kb', 10000],
+      ['1 MB', 1000000],
+      ['100 MB', 100000000],
+      ['10 GB', 10000000000],
+      ['1 TB', 1000000000000],
+      ['100 TB', 100000000000000],
+      ['10 PB', 10000000000000000],
+      ['1 EB', 1000000000000000000],
+      ['100 EB', 100000000000000000000],
+      ['10 ZB', 10000000000000000000000],
+      ['1 YB', 1000000000000000000000000],
+    ],
+  },
+);
